+++ docs/joss/paper.bib
@@ -30,7 +30,7 @@ journal = {Journal of Phycology},
 volume = {9},
 number = {3},
 pages = {264-272},
-doi = {https://doi.org/10.1111/j.1529-8817.1973.tb04092.x},
+doi = {10.1111/j.1529-8817.1973.tb04092.x},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1529-8817.1973.tb04092.x},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1529-8817.1973.tb04092.x},
 year = {1973}
@@ -42,7 +42,8 @@ year = {1973}
   journal={Marine Ecology Progress Series},
   volume={148},
   pages={187--200},
-  year={1997}
+  year={1997},
+  doi={10.3354/meps148187}
 }
 
 @article{wu2021modeling,
@@ -53,7 +54,8 @@ year = {1973}
   number={9},
   pages={e2021GB006941},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2021gb006941}
 }
 
 @misc{Agents2021,
@@ -62,7 +64,8 @@ year = {1973}
       year={2021},
       eprint={2101.10072},
       archivePrefix={arXiv},
-      primaryClass={cs.MA}
+      primaryClass={cs.MA},
+      doi={10.1177/00375497211068820}
 }
 
 @misc{NetLogo1999,
@@ -96,7 +99,8 @@ year = {1973}
   number={7},
   pages={461--471},
   year={2016},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nrmicro.2016.62}
 }
 
 @article{hellweger2014biogeographic,
@@ -107,7 +111,8 @@ year = {1973}
   number={6202},
   pages={1346--1349},
   year={2014},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.1254421}
 }
 
 @article{clark2011individual,
@@ -118,7 +123,8 @@ year = {1973}
   number={23-24},
   pages={3823--3837},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ecolmodel.2011.10.001}
 }
 
 @article{follows2007emergent,
@@ -129,5 +135,6 @@ year = {1973}
   number={5820},
   pages={1843--1846},
   year={2007},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.1138544}
 }
