### 2022-02-17 #4176 @zhenwu0728
potential reviewers:

vchuravy, an expert of CUDA.jl which PlanktonIndividuals.jl based on;
ali-ramadhan,  an expert of ocean modeling and Julia;
simonbyrne, an expert of climate modeling and Julia.
