### 2022-02-17 #4176 @elbeejay
@zhenwu0728 some minor suggestions for the paper:
- Put parentheses around the figure references (line 13, 24, 62)
- Pluralize "waterbody" on line 36 to be consistent with the plural "grids"
- Switch "taylored" to "tailored" on line 53

Overall the submission package looks good to me, and I'm excited to see this be reviewed and eventually published in JOSS.
