### 2022-02-17 #4176 @elbeejay
:wave: @simonbyrne, @navidcy, @Balinus, @simone-silvestri would you be willing to review this submission to JOSS? All of you have backgrounds working in Julia, as well as on ocean, climate, and hydrodynamic models which I believe would allow you to provide high-quality and insightful reviews. At JOSS we do open checklist-driven reviews; peer-review criteria can be viewed [here](https://joss.readthedocs.io/en/latest/review_criteria.html).

The submission is _"PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model"_ . This is a pre-review issue which we use to find peer-reviewers. Once 3 reviewers are found, we will officially start the review in a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review but can recommend someone else, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission.

Thank you, feel free to reach out if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time.
