### 2022-02-18 #4176 @elbeejay
> Thanks for considering me. However, the subject matter is outside my expertise. I could suggest Datseris as a potential reviewer. I am not 100% sure though if he has expertise in oceans. I was about to suggest gaelforget but just found out he is the 2nd author smile

Thanks for the input @Balinus - if you are interested you are welcome to review the Julia implementation with the caveat that your expertise is not in ocean sciences, although I leave that up to you. @Datseris, we would welcome your review of this software if you are interested; I believe you are familiar with the JOSS process and very knowledgeable regarding the Julia language, although the subject matter may be a bit out of scope for you as well.

@navidcy - Thanks for the reply! Would be very appreciative of your review; with regards to your availability, we can plan to start the review process at the end of February (or the first few days of March at the latest) to try and give you at least 3 weeks before you take leave to perform your review. Does that sound reasonable to you?
