+++ src/Grids/lat_lon_grid.jl
@@ -1,7 +1,3 @@
-"""
-    LatLonGrid{TX, TY, TZ} <: AbstractGrid{TX, TY, TZ}
-A horizontally curvilinear grid with topology {TX, TY, TZ}
-"""
 struct LatLonGrid{TX, TY, TZ, R, A1, A2, A3} <: AbstractGrid{TX, TY, TZ}
     # corrdinates at cell centers, unit: degree
     xC::R
