### 2022-02-25 #4207 @editorialbot
[REVIEW]: PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model.
**Submitting author:** @zhenwu0728 (<a href="http://orcid.org/0000-0001-8474-4274">Zhen Wu</a>)
**Repository:** https://github.com/JuliaOcean/PlanktonIndividuals.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.1
