### 2022-02-25 #4207 @elbeejay
@navidcy and @Datseris thanks again for agreeing to review this submission to JOSS. We've just switched over from whedon to the new editorialbot, and the review process has changed subtly. Instead of your checklist being at the top of this issue, you can generate your own checklist with the command `@editorialbot generate my checklist`. Otherwise the review process is functionally the same as it has been in the past -- please feel free to reach out to me with any questions. Remember to open issues in the `PlanktonIndividuals.jl` repository that are related to this JOSS review, and "link" them to this review issue by posting this URL (`https://github.com/openjournals/joss-reviews/issues/4207`) in the text of the issue you open.

We're starting this review before I've found a 3rd reviewer out of respect for @navidcy's availability in late March, so as a heads up to @zhenwu0728 and co-authors, I will still be reaching out to and hopefully adding another reviewer here.
