### 2022-03-03 f2ea83e
CompatHelper: bump compat for CUDAKernels to 0.4, (keep existing compat)
