+++ Project.toml
@@ -20,7 +20,7 @@ StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 [compat]
 Adapt = "^3"
 CUDA = "^3"
-CUDAKernels = "^0.3"
+CUDAKernels = "^0.3, 0.4"
 JLD2 = "^0.4"
 KernelAbstractions = "^0.7"
 StructArrays = "^0.6"
