### 2022-03-11 #4207 @Datseris
Hello @elbeejay , I would like to submit my review:

## Summary
PlanktonIndividuals.jl is a Julia package that simulates the lifecycle and transport of plankton in oceanic currents. It has strong performance capabilities utilizing GPU infrastructure. The paper makes a clear case supporting the need for the package. The paper and authors satisfy the official checklist of JOSS. As far as I can tell, the documentation of the package is very good. It includes documentation strings, examples, and clearly states the actual physics used in the models via stating actual equations, something that (unfortunately) is missing from many packages. I was able to install the software and run their [0D example](https://juliaocean.github.io/PlanktonIndividuals.jl/stable/examples/0D_experiment.html) without problems. All in all, I would recommend publication, however, there are some points I would strongly recommend the authors to consider, and address satisfactorily.

## Additions I would recommend
1. The first couple of sentences of either the Summary or the Statement of Need sections should introduce a general audience to the relevance of phytoplankton _in general_. The paper is written with the expectation that the reader already knows why this topic is useful, important, or relevant to do research on. As such, it is not written for a diverse, non-specialist audience as expected by JOSS guidelines. A couple of sentences can fix this. The paper furthermore does not define what "Eulerian" or "Lagrangian" means.
2. Julia has a general purpose software for individual based modelling, called Agents.jl, which the authors cite. The authors however do not discuss what features or functionalities are missing from Agents.jl that requires them to write a new package instead of extending the existing one. (Full disclosure to the editor: I am a developer of Agents.jl. Admittedly, for me it is actually clear why the authors could not use the existing Agents.jl infrastructure, but I do not believe it would be as clear for a general audience).
3. A discussion on the software design and extendability of the package is missing. For example, how does the software simulate ocean currents that transport phytoplantkon? Does it integrate with existing packages such as Oceanigans.jl? How much flexibility is there in the ocean+plankton interplay? Could it use to simulate other "particles" besides plankton? Could it simulate the tranport of non-alive chemicals, that nevertheless perform reactions with other chemicals? If not, why? is there something fundamentally special about planktons that separates them from other physical entities?
4. The test suite of the package could definitely benefit from improvements. For example, the tests do not test individual atomic functions, as a unit test suite should, but rather entire simulation pipelines. The coverage also seems to be low at only 62%.
5. The documentation does not have installation instructions. Given that this is a registered package, it is easy to resolve this simply by adding "This package is registered, use `Pkg.add("PlanktonIndividuals")` to install it." in the homepage.
6. The package GitHub page does not have Community guidelines.

## Minor technical details
* waterbodys -> water bodies
* Citation Droop (1973) has title in all capitals
* The author affiliation is just "MIT, EAPS".
