### 2022-03-14 #4207 @elbeejay
@navidcy I was going to ping you so that hopefully this review doesn't slip through the cracks before you are unavailable - the new JOSS procedure requires you to create your own reviewer checklist with the command `@editorialbot generate my checklist`

Thanks!
