### 2022-03-17 2484849
add MacroMolecularMode
