+++ src/Plankton/Plankton.jl
@@ -38,20 +38,28 @@ end
 include("Advection/Advection.jl")
 include("QuotaMode/QuotaMode.jl")
 include("CarbonMode/CarbonMode.jl")
+include("MacroMolecularMode/MacroMolecularMode.jl")
 include("utils.jl")
 
 using .Advection
 import .Quota
 import .Carbon
+import .MacroMolecular
 
 #####
 ##### some workarounds for function names
 #####
+construct_plankton(arch::Architecture, sp::Int64, params::Dict, maxN, mode::MacroMolecularMode) = 
+    MacroMolecular.construct_plankton(arch::Architecture, sp::Int64, params::Dict, maxN)
+
 construct_plankton(arch::Architecture, sp::Int64, params::Dict, maxN, mode::QuotaMode) = 
     Quota.construct_plankton(arch::Architecture, sp::Int64, params::Dict, maxN)
 
 construct_plankton(arch::Architecture, sp::Int64, params::Dict, maxN, mode::CarbonMode) = 
     Carbon.construct_plankton(arch::Architecture, sp::Int64, params::Dict, maxN)
+
+generate_plankton!(plank, N::Int64, g::AbstractGrid, arch::Architecture, mode::MacroMolecularMode) =
+    MacroMolecular.generate_plankton!(plank, N::Int64, g::AbstractGrid, arch::Architecture)
 
 generate_plankton!(plank, N::Int64, g::AbstractGrid, arch::Architecture, mode::QuotaMode) =
     Quota.generate_plankton!(plank, N::Int64, g::AbstractGrid, arch::Architecture)
@@ -59,14 +67,19 @@ generate_plankton!(plank, N::Int64, g::AbstractGrid, arch::Architecture, mode::Q
 generate_plankton!(plank, N::Int64, g::AbstractGrid, arch::Architecture, mode::CarbonMode) =
     Carbon.generate_plankton!(plank, N::Int64, g::AbstractGrid, arch::Architecture)
 
+plankton_update!(plank, nuts, proc, p, plk, diags_spcs, ΔT, t, arch::Architecture, mode::MacroMolecularMode) =
+    MacroMolecular.plankton_update!(plank, nuts, proc, p, plk, diags_spcs, ΔT, t, arch::Architecture)
+plankton_update!(plank, nuts, proc, p, plk, nothing::Nothing, ΔT, t, arch::Architecture, mode::MacroMolecularMode) =
+    MacroMolecular.plankton_update!(plank, nuts, proc, p, plk, nothing::Nothing, ΔT, t, arch::Architecture)
+
 plankton_update!(plank, nuts, proc, p, plk, diags_spcs, ΔT, t, arch::Architecture, mode::QuotaMode) =
-    Quota.plankton_update!(plank, nuts, proc, p, plk, diags_spcs, ΔT, t, arch::Architecture, mode::QuotaMode)
+    Quota.plankton_update!(plank, nuts, proc, p, plk, diags_spcs, ΔT, t, arch::Architecture)
 plankton_update!(plank, nuts, proc, p, plk, nothing::Nothing, ΔT, t, arch::Architecture, mode::QuotaMode) =
-    Quota.plankton_update!(plank, nuts, proc, p, plk, nothing::Nothing, ΔT, t, arch::Architecture, mode::QuotaMode)
+    Quota.plankton_update!(plank, nuts, proc, p, plk, nothing::Nothing, ΔT, t, arch::Architecture)
 
 plankton_update!(plank, nuts, proc, p, plk, diags_spcs, ΔT, t, arch::Architecture, mode::CarbonMode) =
[…]
