### 2022-03-17 #4207 @navidcy
> @navidcy I was going to ping you so that hopefully this review doesn't slip through the cracks before you are unavailable - the new JOSS procedure requires you to create your own reviewer checklist with the command `@editorialbot generate my checklist`
>
>
>
> Thanks!

Thanks. I haven't forgotten about this.
