### 2022-03-18 1c1cdc9
add write_species_dynamics for MacroMolecularMode
