+++ docs/joss/paper.bib
@@ -25,7 +25,7 @@ DOI = {10.5194/gmd-10-4175-2017}
 
 @article{Droop1973,
 author = {Droop, M. R.},
-title = {SOME THOUGHTS ON NUTRIENT LIMITATION IN ALGAE1},
+title = {Some thoughts on nutrient limitation in algae},
 journal = {Journal of Phycology},
 volume = {9},
 number = {3},
