### 2022-03-18 fb9fe9b
add default parameter for protein, DNA, and RNA synthesis
