### 2022-03-21 632be47
allow simulation starts with model.interation != 1
