+++ src/Plankton/CarbonMode/growth_kernels.jl
@@ -7,16 +7,16 @@
 end
 
 ##### calculate photosynthesis rate (mmolC/individual/second)
-@inline function calc_PS(par, temp, Chl, Bm, Sz, p)
+@inline function calc_PS(par, temp, Chl, Bm, p)
     αI  = par * p.α * p.Φ
-    PCm = p.PCmax * Sz^p.PC_b * tempFunc(temp, p)
+    PCm = p.PCmax * tempFunc(temp, p)
     PS  = PCm * (1.0 - exp(-αI / max(1.0e-10, PCm) * Chl / max(1.0e-10, Bm))) * Bm
     return PS
 end
 
 @kernel function calc_inorganic_uptake_kernel!(plank, nuts, p)
     i = @index(Global)
-    @inbounds plank.PS[i] = calc_PS(nuts.par[i], nuts.T[i], plank.Chl[i], plank.Bm[i], plank.Sz[i], p) * plank.ac[i]
+    @inbounds plank.PS[i] = calc_PS(nuts.par[i], nuts.T[i], plank.Chl[i], plank.Bm[i], p) * plank.ac[i]
 
 end
 
@@ -30,7 +30,7 @@ end
 ##### calculate respiration (mmolC/individual/second)
 @kernel function calc_respir_kernel!(plank, T, p)
     i = @index(Global)
-    @inbounds plank.resp[i] = p.respir_a * plank.Sz[i]^p.respir_b * plank.Bm[i] * tempFunc(T[i], p) * plank.ac[i]
+    @inbounds plank.resp[i] = p.respir_a * plank.Bm[i] * tempFunc(T[i], p) * plank.ac[i]
 end
 function calc_respir!(plank, T, p, arch)
     kernel! = calc_respir_kernel!(device(arch), 256, (size(plank.ac,1)))
