### 2022-03-21 106aec3
add Temp regulation on synthesis of DNA, RNA, and protein
