+++ src/Plankton/MacroMolecularMode/plankton_growth.jl
@@ -19,7 +19,7 @@ function plankton_growth!(plank, nuts, rnd, p, ΔT, t, arch::Architecture)
     ##### check the probabilities every 10 time steps or 1 hour whichever is shorter
     if t%(ΔT*(min(10,3600÷ΔT))) == 0 
         calc_graz_quadratic!(plank, nuts, p.grz_P, arch)
-        calc_mort!(plank, p, arch)
+        # calc_mort!(plank, p, arch)
         ##### Bernouli-like distribution
         plank.dvid .= 0.5 .* (1.0 .- isless.(plank.DNA ./ (p.C_DNA .* p.Nsuper), 2.0))
         get_probability!(plank, rnd, ΔT, arch)
