### 2022-03-25 d8fa034
update equations of biosynthesis (DNA, RNA, protein)
