+++ src/Plankton/MacroMolecularMode/growth_kernels.jl
@@ -7,35 +7,32 @@
 end
 
 ##### calculate photosynthesis rate (mmolC/individual/second)
-@inline function calc_PS(par, temp, Chl, PRO, DNA, RNA, p)
+@inline function calc_PS(par, temp, Chl, PRO, p)
     αI  = par * p.α * p.Φ
     PCm = p.PCmax * tempFunc(temp, p)
-    Bm = functional_C_biomass(PRO, DNA, RNA)
-    PS  = PCm * (1.0 - exp(-αI / max(1.0e-10, PCm) * Chl / max(1.0e-10, Bm))) * Bm
+    PS  = PCm * (1.0 - exp(-αI / max(1.0e-30, PCm) * Chl / max(1.0e-10, PRO))) * PRO
     return PS
 end
 
 ##### calculate nutrient uptake rate (mmolN/individual/second)
-@inline function calc_NP_uptake(NH4, NO3, PO4, temp, CH, NST, PST, PRO, DNA, RNA, p, ac)
-    Bm = functional_C_biomass(PRO, DNA, RNA)
-    C_tot = total_C_biomass(PRO, DNA, RNA, CH)
+@inline function calc_NP_uptake(NH4, NO3, PO4, temp, NST, PST, PRO, DNA, RNA, p, ac)
     N_tot = total_N_biomass(PRO, DNA, RNA, NST, p)
     P_tot = total_P_biomass(DNA, RNA, PST, p)
-    regQN = max(0.0, min(1.0, (p.NSTmax - N_tot / max(1.0e-10, C_tot)) / (p.NSTmax - p.NSTmin)))
-    regQP = max(0.0, min(1.0, (p.PSTmax - P_tot / max(1.0e-10, C_tot)) / (p.PSTmax - p.PSTmin)))
-    VNH4 = p.VNH4max * regQN * NH4/max(1.0e-10, NH4+p.KsatNH4) * tempFunc(temp, p) * Bm * ac
-    VNO3 = p.VNO3max * regQN * NO3/max(1.0e-10, NO3+p.KsatNO3) * tempFunc(temp, p) * Bm * ac
-    VPO4 = p.VPO4max * regQP * PO4/max(1.0e-10, PO4+p.KsatPO4) * tempFunc(temp, p) * Bm * ac
+    regQN = max(0.0, min(1.0, (p.NSTmax - NST / max(1.0e-30, N_tot)) / (p.NSTmax - p.NSTmin)))
+    regQP = max(0.0, min(1.0, (p.PSTmax - PST / max(1.0e-30, P_tot)) / (p.PSTmax - p.PSTmin)))
+    VNH4 = p.VNH4max * regQN * NH4/max(1.0e-10, NH4+p.KsatNH4) * tempFunc(temp, p) * PRO * ac
+    VNO3 = p.VNO3max * regQN * NO3/max(1.0e-10, NO3+p.KsatNO3) * tempFunc(temp, p) * PRO * ac
+    VPO4 = p.VPO4max * regQP * PO4/max(1.0e-10, PO4+p.KsatPO4) * tempFunc(temp, p) * PRO * ac
     return VNH4, VNO3, VPO4
 end
 
 @kernel function calc_inorganic_uptake_kernel!(plank, nuts, p)
     i = @index(Global)
-    @inbounds plank.PS[i] = calc_PS(nuts.par[i], nuts.T[i], plank.Chl[i], plank.PRO[i], plank.DNA[i], plank.RNA[i], p) * plank.ac[i]
+    @inbounds plank.PS[i] = calc_PS(nuts.par[i], nuts.T[i], plank.Chl[i], plank.PRO[i], p) * plank.ac[i]
 
     @inbounds plank.VNH4[i], plank.VNO3[i], plank.VPO4[i] = 
[…]
