### 2022-03-26 9b7a06c
add chl synthesis in diagnostics
