### 2022-04-04 #4207 @BrodiePearson
@elbeejay Here is my review:

Overall this looks like a useful package with good documentation. I particularly like that the package has both stand-alone utility (lab experiments) and an ability to integrate with other software packages used to simulate ocean flows (e.g., Oceananigans). I only have a few very small comments to add beyond the previous review

## Review Comments

- **Functionality documentation**: The paper does not mention whether the model parameters can be varied (i.e. whether the system can be used to simulate something other than the default biogeochemistry). It looks like this is easy to do, as some parameters are changed in the [global example](https://juliaocean.github.io/PlanktonIndividuals.jl/stable/examples/global_ocean_2D_example.html). Could you add a comment in the paper noting that the biogeochemistry can be modified and the possible extent of these modifiable parameters?

- **Quality of Writing & References**: Some minor typos:
  - Line 11: "cell density and stochiometry"
  - Line 22: "The outputs of *PlanktonIndividuals.jl includes" (currently it sounds like it is referring to the dynamical models of the previous sentence)
  - Line 34: "individual cells or plankton patches" (the comma implies cells==plankton patches, which I assume is not the case)
  - Line 45: Remove "for"
  - Figure 3: Is this showing maps of tracer concentration with overlaid individuals (dots) & their tracks up until that point (lines)? Could you add/expand on this description within the abstract?
