+++ CONTRIBUTING.md
@@ -1,0 +1,8 @@
+Thanks for taking the time to contribute.
+
+Contribution to PlanktonIndividuals.jl is welcome in the following ways:
+
+- Reporting bugs and suggestions by using the issue tracker.
+- Modifying the code or documentation with a pull request.
+
+Please open an issue before you create the pull request so that we make sure your idea is aligned with the project goals  and structure.
