### 2022-04-11 f5a5cbe
New thermal dose equation
