+++ test/grid_test.jl
@@ -1,0 +1,164 @@
+using PlanktonIndividuals.Grids
+
+using MeshArrays
+
+function test_rectilinear_grid()
+    grid = RectilinearGrid(size = (4,6,2), x = (0,12), y = (0,12), z = (0,-8), halo = (2,2,2))
+
+    @test grid.Nx == 4
+    @test grid.Ny == 6
+    @test grid.Nz == 2
+
+    @test grid.Δx == 3.0
+    @test grid.Δy == 2.0
+    @test grid.dzC == [4.0,4.0,4.0,4.0,4.0,4.0]
+    @test grid.dzF == [4.0,4.0,4.0,4.0,4.0,4.0]
+
+    @test length(grid.xC) == 4+2*2
+    @test length(grid.yC) == 6+2*2
+    @test length(grid.zC) == 2+2*2
+    @test length(grid.xF) == 4+2*2
+    @test length(grid.yF) == 6+2*2
+    @test length(grid.zF) == 2+2*2
+
+    return nothing
+end
+
+function test_vertically_stretched_rectilinear_grid()
+    zfs = [0.0, -1.0, -3.0, -5.0, -10.0]
+    grid = RectilinearGrid(size = (4,6,4), x = (0,12), y = (0,12), z = zfs, halo = (2,2,2))
+
+    @test grid.Nx == 4
+    @test grid.Ny == 6
+    @test grid.Nz == 4
+
+    @test grid.Δx == 3.0
+    @test grid.Δy == 2.0
+    @test grid.dzC == [1.0,1.0,1.5,2.0,3.5,5.0,5.0,5.0]
+    @test grid.dzF == [1.0,1.0,1.0,2.0,2.0,5.0,5.0,5.0]
+
+    @test grid.zF == [2.0,1.0,-0.0,-1.0,-3.0,-5.0,-10.0,-15.0]
+    @test grid.zC == [1.5,0.5,-0.5,-2.0,-4.0,-7.5,-12.5,-17.5]
+
+    @test length(grid.xC) == 4+2*2
+    @test length(grid.yC) == 6+2*2
+    @test length(grid.zC) == 4+2*2
+    @test length(grid.xF) == 4+2*2
+    @test length(grid.yF) == 6+2*2
+    @test length(grid.zF) == 4+2*2
+
+    return nothing
+end
+
+function test_rectilinear_areas_volumes()
+    zfs = [0.0, -1.0, -3.0, -5.0, -10.0]
+    grid = RectilinearGrid(size = (4,6,4), x = (0,12), y = (0,12), z = zfs, halo = (2,2,2))
+    @test ΔxC(1,1,1,grid) == grid.Δx
+    @test ΔyC(1,1,1,grid) == grid.Δy
+    @test ΔzC(1,1,1,grid) == grid.dzC[1]
+    @test ΔxF(1,1,1,grid) == grid.Δx
+    @test ΔyF(1,1,1,grid) == grid.Δy
+    @test ΔzF(1,1,1,grid) == grid.dzF[1]
+    @test Ax(1,1,1,grid) == grid.Δy*grid.dzF[1]
+    @test Ay(1,1,1,grid) == grid.Δx*grid.dzF[1]
+    @test Az(1,1,1,grid) == grid.Δx*grid.Δy
+    @test volume(1,1,1,grid) == grid.Δx*grid.Δy*grid.dzF[1]
+
+    return nothing
+end
+
+function test_lat_lon_grid()
+    grid = LatLonGrid(size = (360,160,10), lat = (-80,80), lon = (-180,180), z = (0,-20))
+    @test grid.Nx == 360
+    @test grid.Ny == 160
+    @test grid.Nz == 10
+
[…]
