+++ Project.toml
@@ -25,3 +25,13 @@ JLD2 = "^0.4"
 KernelAbstractions = "^0.7, 0.8"
 StructArrays = "^0.6"
 julia = "^1.6"
+
+[extras]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+Coverage = "a2441757-f6aa-5fb2-8edb-039e3f45d037"
+MeshArrays = "cb8c808f-1acf-59a3-9d2b-6e38d009f683"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+
+[targets]
+test = ["BenchmarkTools", "Coverage", "MeshArrays", "Test"]
