### 2022-04-13 #4207 @zhenwu0728
Thanks for the reminder, @elbeejay. And thanks for all the insightful and constructive comments from @Datseris and @BrodiePearson. We've already started the revision of the paper and the package (e.g. add unit tests) and hopefully finish in one week or so. I'll upload a point-to-point response once finished. Thanks again!
