### 2022-04-15 334cabc
remove unnecessary code
