+++ src/Plankton/CarbonMode/plankton_update.jl
@@ -35,29 +35,3 @@ function plankton_update!(plank, nuts, rnd, p, plk, diags_spcs, ΔT, t, arch::Ar
     ##### diagnostic for individual distribution
     diags_proc!(diags_spcs.num, plank.ac, plank.ac, plank.xi, plank.yi, plank.zi, arch)
 end
-
-function plankton_update!(plank, nuts, rnd, p, plk, nothing::Nothing, ΔT, t, arch::Architecture, mode::CarbonMode)
-    plankton_growth!(plank, nuts, rnd, p, ΔT, t, arch)
-
-    calc_consume!(plk.DIC.data, plank, plank.ac, plank.xi, plank.yi, plank.zi, ΔT, arch)
-
-    ##### check the probabilities every 10 time steps or 1 hour whichever is shorter
-    if t%(ΔT*(min(10,3600÷ΔT))) == 0 
-        ##### grazing
-        grazing!(plank, arch, plk, p)
-
-        ###### mortality
-        mortality!(plank, arch, plk, p)
-
-        ##### division
-        ##### check if the number of individuals exceeded
-        dvidnum = dot(plank.dvid, plank.ac)
-        deactive_ind = findall(x -> x == 0.0, plank.ac)
-        if dvidnum > length(deactive_ind)
-            throw(ArgumentError("number of individual exceeds the capacity"))
-        end
-        ##### do not copy inactive individuals
-        plank.dvid .*= plank.ac
-        divide!(plank, deactive_ind, arch)
-    end
-end
