### 2022-04-15 413595f
bug fix in boundary conditions
