+++ src/Fields/apply_bcs.jl
@@ -4,79 +4,88 @@
 
 # a positive west flux is associated with an *increase* in `Gc` near the west boundary.
 # same for south and bottom.
-@inline function apply_west_bc!(Gc, west_flux, j, k, iter, grid, ΔT)
-    @inbounds Gc[1+grid.Hx, j, k] += getbc(west_flux, j, k, iter) * ΔT * Ax(1+grid.Hx, j, k, grid) / volume(1+grid.Hx, j, k, grid)
-    return nothing
+@kernel function apply_west_bcs_kernel!(Gc, grid, west_bc, iter, ΔT)
+    j, k = @index(Global, NTuple)
+    jj = j + grid.Hy
+    kk = k + grid.Hz
+    @inbounds Gc[1+grid.Hx, jj, kk] += getbc(west_bc, j, k, iter) * ΔT * Ax(1+grid.Hx, jj, kk, grid) / volume(1+grid.Hx, jj, kk, grid)
 end
 
-@inline function apply_south_bc!(Gc, south_flux, i, k, iter, grid, ΔT)
-    @inbounds Gc[i, 1+grid.Hy, k] += getbc(south_flux, i, k, iter) * ΔT * Ay(i, 1+grid.Hy, k, grid) / volume(i, 1+grid.Hy, k, grid)
-    return nothing
+@kernel function apply_south_bcs_kernel!(Gc, grid, south_bc, iter, ΔT)
+    i, k = @index(Global, NTuple)
+    ii = i + grid.Hx
+    kk = k + grid.Hz
+    @inbounds Gc[ii, 1+grid.Hy, kk] += getbc(south_bc, i, k, iter) * ΔT * Ay(ii, 1+grid.Hy, kk, grid) / volume(ii, 1+grid.Hy, kk, grid)
 end
 
-@inline function apply_bottom_bc!(Gc, bottom_flux, i, j, iter, grid, ΔT)
-    @inbounds Gc[i, j, grid.Nz+grid.Hz] += getbc(bottom_flux, i, j, iter) * ΔT * Az(i, j, grid.Nz+grid.Hz, grid) / volume(i, j, grid.Nz+grid.Hz, grid)
-    return nothing
+@kernel function apply_bottom_bcs_kernel!(Gc, grid, bottom_bc, iter, ΔT)
+    i, j = @index(Global, NTuple)
+    ii = i + grid.Hx
+    jj = j + grid.Hy
+    @inbounds Gc[ii, jj, grid.Nz+grid.Hz] += getbc(bottom_bc, i, j, iter) * ΔT * Az(ii, jj, grid.Nz+grid.Hz, grid) / volume(ii, jj, grid.Nz+grid.Hz, grid)
 end
 
 # a positive east flux is associated with an *decrease* in `Gc` near the east boundary.
 # same for north and top.
-@inline function apply_east_bc!(Gc, east_flux, j, k, iter, grid, ΔT)
-    @inbounds Gc[grid.Nx+grid.Hx, j, k] += getbc(east_flux, j, k, iter) * ΔT * Ax(grid.Nx+grid.Hx, j, k, grid) / volume(grid.Nx+grid.Hx, j, k, grid)
-    return nothing
-end
-
-@inline function apply_north_bc!(Gc, north_flux, i, k, iter, grid, ΔT)
-    @inbounds Gc[i, grid.Ny+grid.Hy, k] += getbc(north_flux, i, k, iter) * ΔT * Ay(i, grid.Ny+grid.Hy, k, grid) / volume(i, grid.Ny+grid.Hy, k, grid)
-    return nothing
-end
-
-@inline function apply_top_bc!(Gc, top_flux, i, j, iter, grid, ΔT)
[…]
