+++ test/example_2D_test.jl
@@ -41,3 +41,7 @@ TPt = tot_mass(model.nutrients.PO4.data, grid) +
       tot_mass(model.nutrients.POP.data, grid)
 TPt = TPt + sum(model.individuals.phytos.sp1.data.Pq .+ 
                 model.individuals.phytos.sp1.data.Bm .* model.individuals.phytos.sp1.p.R_PC)
+
+@testset "PlanktonIndividuals 2D tests:" begin
+    @test isapprox(TP,TPt; atol=1e1)
+end 
