### 2022-04-15 90eda2a
add unit test for parameters and output
