### 2022-04-15 f8b8fbf
Update write_outputs.jl
