### 2022-04-15 2f1fd97
add test for MacroMolecularMode
