+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"function_index/#Function-Index","page":"Function index","title":"Function Index","text":"","category":"section"},{"location":"function_index/","page":"Function index","title":"Function index","text":"","category":"page"},{"location":"benchmarks/#benchmarks","page":"Benchmarks","title":"Benchmarks","text":"","category":"section"},{"location":"benchmarks/","page":"Benchmarks","title":"Benchmarks","text":"Here we benchmark the model performance in two Architectures. The number of individuals used in the benchmark are (2^5, 2^10, 2^15, 2^17). And we also use different grid resolutions in 2-Dimensional and 3-Dimensional model setup.","category":"page"},{"location":"benchmarks/#Dimensional-model","page":"Benchmarks","title":"0-Dimensional model","text":"","category":"section"},{"location":"benchmarks/","page":"Benchmarks","title":"Benchmarks","text":"This is a benchmark of a simple 0-Dimensional model setup without advection of Eulerian tracers. However, the advection of individuals still take the same amount of time whether the velocity field is provided or not.","category":"page"},{"location":"benchmarks/","page":"Benchmarks","title":"Benchmarks","text":"PlanktonIndividuals v0.4.2\nJulia Version 1.7.0-rc1\nCommit 9eade6195e (2021-09-12 06:45 UTC)\nPlatform Info:\n  OS: Linux (x86_64-pc-linux-gnu)\n  CPU: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz\n  WORD_SIZE: 64\n  LIBM: libopenlibm\n  LLVM: libLLVM-12.0.1 (ORCJIT, broadwell)\n  GPU: Tesla P100-PCIE-12GB","category":"page"},{"location":"benchmarks/","page":"Benchmarks","title":"Benchmarks","text":"Arch N min median mean max memory allocs\nCPU 32 978.736 μs 1.062 ms 1.114 ms 1.745 ms 639.39 KiB 3377\nCPU 1024 3.217 ms 3.319 ms 3.357 ms 4.003 ms 639.39 KiB 3377\nCPU 32768 73.551 ms 73.612 ms 73.955 ms 77.018 ms 638.91 KiB 3346\nCPU 131072 297.726 ms 298.756 ms 300.489 ms 316.688 ms 638.91 KiB 3346\nGPU 32 7.498 ms 7.566 ms 7.636 ms 8.331 ms 2.27 MiB 16453\nGPU 1024 7.599 ms 7.691 ms 7.755 ms 8.487 ms 2.26 MiB 16443\nGPU 32768 8.171 ms 8.362 ms 8.470 ms 9.745 ms 2.26 MiB 16443\nGPU 131072 9.698 ms 10.456 ms 10.637 ms 12.999 ms 2.26 MiB 16438","category":"page"},{"location":"benchmarks/#Dimensional-model-2","page":"Benchmarks","title":"2-Dimensional model","text":"","category":"section"},{"location":"benchmarks/","page":"Benchmarks","title":"Benchmarks","text":"This is the benchmark of a 2-Dimensional model setup with (Ns, 1, Ns) grid cells. Here Ns = [32, 64, 128].","category":"page"},{"location":"benchmarks/","page":"Benchmarks","title":"Benchmarks","text":"PlanktonIndividuals v0.4.2\nJulia Version 1.7.0-rc1\nCommit 9eade6195e (2021-09-12 06:45 UTC)\nPlatform Info:\n  OS: Linux (x86_64-pc-linux-gnu)\n  CPU: I
[…]
