### 2022-04-21 #4207 @zhenwu0728
Hi @elbeejay, @Datseris, and @BrodiePearson,
Thanks for all your comments and suggestions.
With the help of @gaelforget, we would like to submit our responses:
Please note that the revised package and paper are tagged v0.5.0.

# Reply to @Datseris

## Summary
PlanktonIndividuals.jl is a Julia package that simulates the lifecycle and transport of plankton in oceanic currents. It has strong performance capabilities utilizing GPU infrastructure. The paper makes a clear case supporting the need for the package. The paper and authors satisfy the official checklist of JOSS. As far as I can tell, the documentation of the package is very good. It includes documentation strings, examples, and clearly states the actual physics used in the models via stating actual equations, something that (unfortunately) is missing from many packages. I was able to install the software and run their [0D example](https://juliaocean.github.io/PlanktonIndividuals.jl/stable/examples/0D_experiment.html) without problems. All in all, I would recommend publication, however, there are some points I would strongly recommend the authors to consider, and address satisfactorily.

**Reply**: Thank you for the positive evaluation of our package and paper.

## Additions I would recommend
1. The first couple of sentences of either the Summary or the Statement of Need sections should introduce a general audience to the relevance of phytoplankton _in general_. The paper is written with the expectation that the reader already knows why this topic is useful, important, or relevant to do research on. As such, it is not written for a diverse, non-specialist audience as expected by JOSS guidelines. A couple of sentences can fix this. The paper furthermore does not define what "Eulerian" or "Lagrangian" means.

    **Reply**: Thanks for pointing it out. We have added two sentences at the beginning of Summary. Also, "Eulerian" or "Lagrangian" are explained at the end of the second paragraph of Summary.

2. Julia has a general purpose software for individual based modelling, called Agents.jl, which the authors cite. The authors however do not discuss what features or functionalities are missing from Agents.jl that requires them to write a new package instead of extending the existing one. (Full disclosure to the editor: I am a developer of Agents.jl. Admittedly, for me it is actually clear why the authors could not use the existing Agents.jl infrastructure, but I do not believe it would be as clear for a general audience).

    **Reply**: Thanks for the comment. We are sorry that we did not state clearly in the paper about the purpose of our package and how it compares with other individual based models. Now we have added some text in the fourth and fifth paragraph in Statement of Need for clarification.

3. A discussion on the software design and extendability of the package is missing. For example, how does the software simulate ocean currents that transport phytoplantkon? Does it integrate with existing packages such as Oceanigans.jl? How much flexibility is there in the ocean+plankton interplay? Could it simulate other "particles" besides plankton? Could it simulate the tranport of non-alive chemicals, that nevertheless perform reactions with other chemicals? If not, why? is there something fundamentally special about planktons that separates them from other physical entities? It really is not clear to me why this package is named "PlanktonIndividuals" instead of something like "InvidualOceanTranport" or so. Given my experience in individual based modelling, I do not see a particularly reason of why the infastructure of this package would be limited only to Plankton.

    **Reply**: Thank you for the comment. Individual-based modeling and the transport of Lagrangian particles are part of the framework of our package. However, its defining component is the phytoplankton physiological equations and biogeochemical equations embeded in this framework. The possibility to simplify equations in the model to simulate simple passive tracers and particles is shown in Fig 3. Yet, representing all types of material interactions within the Ocean is beyond the scope of our package.

    The discussion of the package design has been extended in the last two paragraphs of the summary. Possible extensions to the model are provided in the final paragraph of the paper. Others can be envisioned but those listed are viewed as rather high priority.

4. The test suite of the package could definitely benefit from improvements. For example, the tests do not test individual atomic functions, as a unit test suite should, but rather entire simulation pipelines. The coverage also seems to be low at only 62%.

    **Reply**: Thank you for the suggestion. We have now added unit tests and brought up the coverage to 80%.

5. The documentation does not have installation instructions. Given that this is a registered package, it is easy to resolve this simply by adding "This package is registered, use `Pkg.add("PlanktonIndividuals")` to install it." in the homepage.

    **Reply**: It has been added in Readme.

6. The package GitHub page does not have Community guidelines.

    **Reply**: Thanks for pointing it out. We have added a CONTRIBUTING.md in the package.

## Minor technical details
* waterbodys -> water bodies
* Citation Droop (1973) has title in all capitals
* The author affiliation is just "MIT, EAPS".

    **Reply**: These have been revised in the paper.

# Reply to @BrodiePearson

Overall this looks like a useful package with good documentation. I particularly like that the package has both stand-alone utility (lab experiments) and an ability to integrate with other software packages used to simulate ocean flows (e.g., Oceananigans). I only have a few very small comments to add beyond the previous review

## Review Comments
* **Functionality documentation**: The paper does not mention whether the model parameters can be varied (i.e. whether the system can be used to simulate something other than the default biogeochemistry). It looks like this is easy to do, as some parameters are changed in the [global example](https://juliaocean.github.io/PlanktonIndividuals.jl/stable/examples/global_ocean_2D_example.html). Could you add a comment in the paper noting that the biogeochemistry can be modified and the possible extent of these modifiable parameters?

    **Reply**: Thanks for the comment. A sentence about the modifiable parameters is added in the third paragraph of Summary.

* **Quality of Writing & References**: Some minor typos:
    * Line 11: "cell density and stochiometry"
    * Line 22: "The outputs of *PlanktonIndividuals.jl includes" (currently it sounds like it is referring to the dynamical models of the previous sentence)
    * Line 34: "individual cells or plankton patches" (the comma implies cells==plankton patches, which I assume is not the case)
    * Line 45: Remove "for"
    * Figure 3: Is this showing maps of tracer concentration with overlaid individuals (dots) & their tracks up until that point (lines)? Could you add/expand on this description within the abstract?

        **Reply**: These have been revised in the paper.
