+++ v0.5.0/examples/horizontal_2D_example.jl
@@ -1,0 +1,1216 @@
+### A Pluto.jl notebook ###
+# v0.17.7
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 54683b9a-bf00-4956-8894-e263eded3db8
+using PlanktonIndividuals, Plots, PlutoUI
+
+# ╔═╡ e74d8d84-546b-4ccd-9724-78ed29f587c3
+begin
+	p=dirname(pathof(PlanktonIndividuals))
+	include(joinpath(p,"../examples/helper_functions.jl"))
+	nothing
+end
+
+# ╔═╡ a1e895f9-fd5a-4908-a760-344b7f8fb304
+md"""
+# Horizontal 2-Dimensional Example
+Here we simulate phytoplankton cells as Lagrangian particles in a horizontal, two-dimensional, flow field.
+The domain is periodic in both directions.
+"""
+
+# ╔═╡ a6072dc8-2d64-11ec-0a4e-dd442a0f0e63
+md"""
+## 1. Import packages
+"""
+
+# ╔═╡ 88c6027d-bc47-4bc4-9628-c946ac2fd57f
+md"""
+## 2. Generate Flow Fields
+First we generate grid information and the computational architecture (CPU).
+"""
+
+# ╔═╡ 10d290fd-5aa7-4b7f-be91-3cc51e4250be
+arch = CPU()
+
+# ╔═╡ 3093a948-630d-402c-aeb4-73a552e0cb31
+grid = RectilinearGrid(size=(128, 128, 1), x = (0.0,128meters), y = (0.0,128meters), z = (0.0,-1meter))
+
+# ╔═╡ 02a7cda3-4a3a-49ea-a646-e1da7a1e1f14
+md"""
+Then we use a stream function to generate the flow field which is a double-gyre configuration [as explained here](https://shaddenlab.berkeley.edu/uploads/LCS-tutorial/examples.html#Sec7.1)
+"""
+
+# ╔═╡ ed5b3467-8bac-4800-824f-2bf26b3954c0
+(uvels, vvels, wvels, ϕcenters) = streamfunction_xy();
+
+# ╔═╡ 5abbed9f-a256-4312-a9de-7277379bb33a
+md"""
+## 3. Model Setup
+
+Next we setup the individual-based model by specifying the architecture, grid, and plankton community.
+"""
+
+# ╔═╡ 7735454f-bc9f-4dad-8c40-2e5bba096307
+model = PlanktonModel(arch, grid; N_species = 1, 
+								  N_individual = 2^7,
+								  max_individuals = 2^7*8)
+
+# ╔═╡ 1ec4dee6-e560-4bf1-b523-134961234617
+md"""
+We modify the initial tracer field to highlight advection.
+
+!!! note
+    DOC is used here as a generic tracer; not to represent dissolved organic carbon per se.
+"""
+
+# ╔═╡ 8760c506-bd00-47cd-9560-d23141797953
+for i in 1:132
+	for j in 1:132
+		model.nutrients.DOC.data[i,j,:] .= isless(132-j,i) .+ 1
+	end
+end
+
+# ╔═╡ 970cba93-d4a5-4579-a436-465085ea6e6b
+md"""
+Finally we setup the duration of the model simulation and the kind of output we want.
+"""
+
[…]
