+++ v0.5.0/examples/global_ocean_3D_example.jl
@@ -1,0 +1,1460 @@
+### A Pluto.jl notebook ###
+# v0.17.7
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 88e0998b-177d-4de6-8ebf-932a6969ee7a
+using PlanktonIndividuals
+
+# ╔═╡ 1251276b-991b-4c5c-af05-697f439d8dad
+md"""# Global Ocean 3D example
+
+Here we setup a passive tracer and particle simulation. The tracer is initialized to 1 below 100m depth and 0 above that depth horizon to highlight the Eulerian advection capability. 3600 particles are initialy distributed at random over the Global Ocean to demonstrate the Lagrangian advection capability.
+
+!!! note
+    This simulation may take several hours to complete a one-year run. Here we only run a few time steps.
+
+"""
+
+# ╔═╡ 111b70b4-4bca-11ec-30fa-6318c8a55e2c
+import MeshArrays, OceanStateEstimation, NetCDF,Plots
+
+# ╔═╡ 1431ecc6-2b02-4c86-bd81-906899e78f18
+begin
+	MeshArrays.GRID_LL360_download();
+	γ = MeshArrays.GridSpec("PeriodicChannel",MeshArrays.GRID_LL360);
+	Γ = MeshArrays.GridLoad(γ;option="full");
+	
+	grid_info=(XC=Γ.XC[1], YC=Γ.YC[1], RC=Γ.RC,
+        	XW=Γ.XW[1], YS=Γ.YS[1], RF=Γ.RF,
+            DXC=Γ.DXC[1], DYC=Γ.DYC[1], DRC=Γ.DRC,
+            DXG=Γ.DXG[1], DYG=Γ.DYG[1], DRF=Γ.DRF,
+            RAW=Γ.RAW[1], RAS=Γ.RAS[1], RAC=Γ.RAC[1],
+            hFacC=ones(360,160,50), hFacW=ones(360,160,50), hFacS=ones(360,160,50));
+	
+	OceanStateEstimation.get_occa_variable_if_needed("DDtheta")
+	OceanStateEstimation.get_occa_variable_if_needed("DDuvel")
+	OceanStateEstimation.get_occa_variable_if_needed("DDvvel")
+	OceanStateEstimation.get_occa_variable_if_needed("DDwvel")
+
+	pth=OceanStateEstimation.OCCAclim_path
+end
+
+# ╔═╡ 4efcd71d-2fef-4f80-b8fe-5fc22a2ac056
+function rd(filename, varname, month)
+    fil = NetCDF.open(filename, varname)
+    tmp = fil[:,:,:,month]
+    tmp[findall(tmp.<-1e22)] .= 0.0
+    return tmp
+end
+
+# ╔═╡ e3a11fe6-da49-499e-afe2-91cde6866681
+begin
+	uvel = []; vvel = []; wvel = []; temp = [];
+	for i in 1:12
+		U = zeros(360, 160, 50)
+		V = zeros(360, 161, 50)
+		W = zeros(360, 160, 51)
+		U[:,:,:] .= rd(joinpath(pth,"DDuvel.0406clim.nc"),"u",i)
+		V[:,1:160,:] .= rd(joinpath(pth,"DDvvel.0406clim.nc"),"v",i)
+		W[:,:,2:51] .= rd(joinpath(pth,"DDwvel.0406clim.nc"),"w",i)
+		T=rd(joinpath(pth,"DDtheta.0406clim.nc"),"theta",i)
+		
[…]
