### 2022-04-26 #4207 @BrodiePearson
@elbeejay @zhenwu0728 Thanks for making the modifications. This looks good to me now. (note, there is a minor typo in the new Figure 3 caption: "liens" to "lines")
