### 2022-04-29 #4207 @elbeejay
@zhenwu0728 and @gaelforget I have some minor suggestions for edits I think you should make to the paper.

- Remove the period from the title
- Either define both GPU and CPU when they are first introduced, or define neither. As currently written only GPU is defined
- Line 63: I believe the part of the sentence "both in Julia or other" should be "both in Julia *and* other" swapping "or" for "and"
- Please select a single convention for referring to dimensionality and use it throughout the paper, I think upwards of three different variations are currently used (three-dimensional, 3D, 3-dimensional)
- Line 76: "have a two-way interactions", I think the "a" can be deleted
- In the caption for Figure 1, add the ` marks around "PlanktonIndividuals.jl" so it is typeset consistently with the rest of the paper
- In Lines 45-46 of the statement of need there should only be a single set of parentheses capturing both the e.g., statement and the references. The syntax for this should be: `[e.g., lakes, ..., @citeA; @citeB]`

Minor comments aside this seems like it's about ready to go. Once the the above changes to the paper are made (and I would suggest you both do a final read through yourselves to ensure it reads the way you'd like), there are a few final archival steps I need you to do:

Please let me know when these steps are complete, they allow us to finalize the metadata that will be associated with the JOSS publication. Once that is done I can recommend that we accept and publish this paper.
