+++ docs/joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model.'
+title: 'PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model'
 tags:
   - Julia
   - Individual-based model
@@ -29,28 +29,28 @@ Marine phytoplankton contributes ~50% of the total primary production on Earth f
 
 `PlanktonIndividuals.jl` can simulate multiple functional groups of phytoplankton with different growth and division strategies. It can thus effectively represent the interactions within and between functional groups. It simulates phytoplankton cells as Lagrangian particles (moving frame) while nutrients are represented as Eulerian tracers (fixed frame). The model thus takes advantage of both frameworks -- Eulerian simulations employ a fixed mesh to represent flow fields on a grid, while plankton individuals move following the velocity field interpolated to their locations.
 
-![Schematic diagram of phytoplankton physiology described in PlanktonIndividuals.jl.\label{fig:phyto}](PI_Quota.jpeg)
+![Schematic diagram of phytoplankton physiology described in `PlanktonIndividuals.jl`.\label{fig:phyto}](PI_Quota.jpeg)
 
 The equations included in `PlanktonIndividuals.jl` to describe phytoplankton physiology are state of the art (\autoref{fig:phyto}) with the widely-used Droop model [@Droop1973] implemented for nutrient uptakes. The photosynthesis formulation by Geider et al [@geider1997dynamic] is also implemented for carbon fixation. Additionally, exudation and mixotrophy are also accounted for in the model [@wu2021modeling]. Both simulated phytoplankton cells and nutrient tracers are advected by the same velocity field provided externally. The growth rate and nutrient uptake rates of each simulated cell are decided by the nutrient availability, light intensity, and ambient temperature at the position of the cell. And in turn, the nutrient consumed or released by phytoplankton cells are added to (or subtracted from) the nutrient fields at the grid point the cells are in. 
 
 The simulations can be set up not only in a zero-dimensional domain (like lab experiments) but also in one-, two-, or three-dimensional domains. The external inputs required by the package include velocity fields, temperature fields, and surface photosynthetically active radiation (PAR) fields.  Our package is able to use velocity and temperature fields provided by different fluid dynamic models (e.g. `Oceananigans.jl`, `MITgcm`, etc.) and observations. Two default parameter sets are inclued in the package, biogeochemical parameters of nutrient tracers, and phytoplankton physiological parameters for individual phytoplankton cells. These parameters can be varied to simulate different phytoplankton species and biogeochemistry. The outputs of `PlanktonIndividuals.jl` include time series of individual trajectories, cell density, biomass, and vital rates such as cell division rates, photosynthesis rate, and nutrient uptake rates (\autoref{fig:0D}).
 
-![Model results of a 0-dimensional setup.\label{fig:0D}](0D_plot.png)
+![Model results of a zero-dimensional setup.\label{fig:0D}](0D_plot.png)
 
 # Statement of need
-Phytoplankton models, like those incorporated in ecological and biogeochemical models, are essential tools to explore and understand phytoplankton life cycles, interactions between species, population dynamics, and community structures in various water bodies (e.g. lakes, coastal waters, and open ocean) [@follows2007emergent; @geider1997dynamic]. Individual-based / Lagrangian models differ from Eulerian models in the way they represent heterogeneous populations like cells of different sizes, nutrient quotas, and life cycle stages [@hellweger2016advancing; @hellweger2014biogeographic; @clark2011individual]. 
[…]
