+++ Project.toml
@@ -1,7 +1,7 @@
 name = "PlanktonIndividuals"
 uuid = "dfeb4ede-c75e-11e9-050d-41ab7a730e9e"
 authors = ["zhenwu0728 <zhenwu@mit.edu>", "gaelforget <gforget@mit.edu>"]
-version = "0.5.0"
+version = "0.5.1"
 
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
