### 2022-05-01 #4207 @elbeejay
Hi @zhenwu0728,

Thanks for making those revisions. One minor thing, and I apologize if I was unclear above, please revise the citation in the Statement of need to be: `[e.g. lakes, coastal waters, and open ocean, @follows2007emergent; @geider1997dynamic]`, in the `paper.md` file, which I believe should end up getting typeset as: (e.g. lakes, coastal waters, and open ocean, Follows et al., 2007; Geider et al. 1997) in the compiled PDF. Let me know if this doesn't work, my intent is for the paper to no longer have the nested parentheses there.
