### 2022-05-04 #4207 @kyleniemeyer
Hi @zhenwu0728, I'm the AEIC on duty doing some final checks before publishing.

In your article, could you

Also, It looks like there is actually a JOSS article on Agents.jl you should probably cite (https://doi.org/10.21105/joss.01611), perhaps in addition to the newer article you do cite.
