+++ docs/joss/paper.bib
@@ -58,14 +58,31 @@ year = {1973}
   doi={10.1029/2021gb006941}
 }
 
-@misc{Agents2021,
-      title={Agents.jl: A performant and feature-full agent based modelling software of minimal code complexity}, 
-      author={George Datseris and Ali R. Vahdati and Timothy C. DuBois},
-      year={2021},
-      eprint={2101.10072},
-      archivePrefix={arXiv},
-      primaryClass={cs.MA},
-      doi={10.1177/00375497211068820}
+@article{Agents2022,
+  doi = {10.1177/00375497211068820},
+  url = {https://doi.org/10.1177/00375497211068820},
+  year = {2022},
+  month = jan,
+  publisher = {{SAGE} Publications},
+  pages = {003754972110688},
+  author = {George Datseris and Ali R. Vahdati and Timothy C. DuBois},
+  title = {Agents.jl: a performant and feature-full agent-based modeling software of minimal code complexity},
+  journal = {{SIMULATION}},
+  volume = {0},
+  number = {0},
+}
+
+@article{Agents2019,
+  doi = {10.21105/joss.01611},
+  url = {https://doi.org/10.21105/joss.01611},
+  year = {2019},
+  publisher = {The Open Journal},
+  volume = {4},
+  number = {42},
+  pages = {1611},
+  author = {Ali R. Vahdati},
+  title = {Agents.jl: agent-based modeling framework in Julia},
+  journal = {Journal of Open Source Software}
 }
 
 @misc{NetLogo1999,
